Added type hinting, new format, code refactoring, transporters can have their own level specified.
If the logger is assigned a name when creating the PyLogger the LoggerName format will write the name of the logger in its output. 
Transporters can have their own 'Level' specified and define if it has to be the same as the log level or includes all of the above of the log level.